Add api::v1 namespace to chats and documents endpoints
