+++ config/routes.rb
@@ -12,10 +12,14 @@ Rails.application.routes.draw do
   get "service-worker" => "rails/pwa#service_worker", as: :pwa_service_worker
   get "manifest" => "rails/pwa#manifest", as: :pwa_manifest
 
-  # Chat routes
-  root "chats#index"
-  post "chat", to: "chats#chat"
+  namespace :api, defaults: { format: :json } do
+    namespace :v1 do
+      # Chat routes
+      root "chats#index"
+      post "chat", to: "chats#chat"
 
-  # Document routes
-  resources :documents
+      # Document routes
+      resources :documents
+    end
+  end
 end
